+++ src/stacktrace_generic-inl.h
@@ -38,10 +38,10 @@
 _START_GOOGLE_NAMESPACE_
 
 // If you change this function, also change GetStackFrames below.
-size_t GetStackTrace(void** result, size_t max_depth, int skip_count) {
+size_t GetStackTrace(void** result, size_t max_depth, size_t skip_count) {
   static const int kStackLength = 64;
   void * stack[kStackLength];
-  int size;
+  size_t size;
 
   size = backtrace(stack, kStackLength);
   skip_count++;  // we want to skip the current frame as well
@@ -50,7 +50,7 @@ size_t GetStackTrace(void** result, size_t max_depth, int skip_count) {
     result_count = 0;
   if (result_count > max_depth)
     result_count = max_depth;
-  for (int i = 0; i < result_count; i++)
+  for (size_t i = 0; i < result_count; i++)
     result[i] = stack[i + skip_count];
 
   return result_count;
